+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/tobe/DispatcherServletTest.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc.tobe;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
